moved contestScrapper.js to /scrapper
